requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-taggit==3.0.0
gunicorn==20.1.0
Markdown==3.4.1
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
wheel==0.37.1

